Fix mobile footer waveform display
The footer waveform on small screens (possibly just mobile safari?)
displays the whole graphic, instead of just the top of the waveform as
desired.

Also, the graphic is shrunken down to be the size to the view port
width, making the effect harder to track.

Cut off overflowing parts of the waveform instead of shrinking the
entire graphic to fit, which keeps the effect readable.

Add a cute hover effect for each beam.