Fix basic artists listing page
Remove the hard coded stuff in the index.php file; that's breaking other
pages.

Pass in the query vars to the function, so it doesn't zero them
out; this still doesn't do anything for us, but it I think it will be
required for getting artist individual pages up and running.

In production, there is a page I created for the artists listing, with
the ID of 378. Use that reference to hook into for the custom code.
There might be another way of doing it purely in code, but God if I've
wasted enough time mucking around with Wordpress re-write rules trying
to figure it out, and I'm just done.